# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Spendly API
  version: 1.0.0
servers:
  - url: https://api.spendly.app
paths:
  /v1/groups:
    get:
      operationId: listGroups
      security: [{ bearerAuth: [] }]
      responses: { '200': { description: Groups } }
    post:
      operationId: createGroup
      security: [{ bearerAuth: [] }]
      responses: { '201': { description: Group created } }
  /v1/groups/{id}:
    patch:
      operationId: archiveGroup
      security: [{ bearerAuth: [] }]
      responses: { '204': { description: Group archived } }
  /v1/groups/{id}/members/{userID}:
    patch:
      operationId: updateGroupMember
      security: [{ bearerAuth: [] }]
      responses: { '204': { description: Member updated } }
  /v1/groups/{id}/invitations:
    post:
      operationId: createGroupInvitation
      security: [{ bearerAuth: [] }]
      responses: { '201': { description: Invitation created } }
  /v1/group-invitations/{token}/accept:
    post:
      operationId: acceptGroupInvitation
      security: [{ bearerAuth: [] }]
      responses: { '204': { description: Invitation accepted } }
  /v1/purchases:
    get:
      operationId: listPurchases
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: from, in: query, required: true, schema: { type: string, format: date-time } }
        - { name: to, in: query, required: true, schema: { type: string, format: date-time } }
        - { name: group_id, in: query, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Purchases }
        '401': { $ref: '#/components/responses/Error' }
    post:
      operationId: createPurchase
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: Idempotency-Key, in: header, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PurchaseDraft' }
      responses:
        '201': { description: Purchase created }
        '409': { $ref: '#/components/responses/Error' }
  /v1/purchases/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      - { name: If-Match, in: header, required: true, schema: { type: integer, format: int64 } }
    put:
      operationId: updatePurchase
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Purchase updated }
        '409': { $ref: '#/components/responses/Error' }
    delete:
      operationId: deletePurchase
      security: [{ bearerAuth: [] }]
      responses:
        '204': { description: Purchase deleted }
        '409': { $ref: '#/components/responses/Error' }
  /v1/auth/google:
    post:
      operationId: signInWithGoogle
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProviderSignInRequest' }
      responses:
        '200': { $ref: '#/components/responses/Session' }
        '401': { $ref: '#/components/responses/Error' }
        '503': { $ref: '#/components/responses/Error' }
  /v1/auth/apple:
    post:
      operationId: signInWithApple
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProviderSignInRequest' }
      responses:
        '200': { $ref: '#/components/responses/Session' }
        '401': { $ref: '#/components/responses/Error' }
        '503': { $ref: '#/components/responses/Error' }
  /v1/auth/refresh:
    post:
      operationId: refreshSession
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RefreshRequest' }
      responses:
        '200': { $ref: '#/components/responses/Session' }
        '401': { $ref: '#/components/responses/Error' }
  /v1/auth/logout:
    post:
      operationId: logout
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RefreshRequest' }
      responses:
        '204': { description: Session revoked }
  /v1/auth/logout-all:
    post:
      operationId: logoutAll
      security: [{ bearerAuth: [] }]
      responses:
        '204': { description: All sessions revoked }
        '401': { $ref: '#/components/responses/Error' }
  /v1/me:
    get:
      operationId: getCurrentUser
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                type: object
                required: [user]
                properties:
                  user: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Error' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    PurchaseDraft:
      type: object
      additionalProperties: false
      required: [id, kind, merchant, currency_code, spent_at, local_date, time_zone]
      properties:
        id: { type: string, format: uuid }
        group_id: { type: string, format: uuid }
        kind: { type: string, enum: [quick, detailed] }
        merchant: { type: string, minLength: 1 }
        category: { type: string }
        amount_minor: { type: integer, format: int64 }
        currency_code: { type: string, pattern: '^[A-Z]{3}$' }
        spent_at: { type: string, format: date-time }
        local_date: { type: string, format: date }
        time_zone: { type: string, minLength: 1 }
        items:
          type: array
          items: { $ref: '#/components/schemas/PurchaseItem' }
    PurchaseItem:
      type: object
      required: [id, position, name, category, amount_minor]
      properties:
        id: { type: string, format: uuid }
        position: { type: integer, minimum: 0 }
        name: { type: string, minLength: 1 }
        category: { type: string, minLength: 1 }
        amount_minor: { type: integer, format: int64, minimum: 1 }
    ProviderSignInRequest:
      type: object
      additionalProperties: false
      required: [id_token, nonce]
      properties:
        id_token: { type: string, minLength: 1 }
        nonce: { type: string, minLength: 1 }
    RefreshRequest:
      type: object
      additionalProperties: false
      required: [refresh_token]
      properties:
        refresh_token: { type: string, minLength: 1 }
    User:
      type: object
      required: [id]
      properties:
        id: { type: string, format: uuid }
        display_name: { type: string }
        currency_code: { type: string }
        time_zone: { type: string }
    Session:
      type: object
      required: [access_token, refresh_token, access_expires_at, refresh_expires_at]
      properties:
        user: { $ref: '#/components/schemas/User' }
        access_token: { type: string }
        refresh_token: { type: string }
        access_expires_at: { type: string, format: date-time }
        refresh_expires_at: { type: string, format: date-time }
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code: { type: string }
            message: { type: string }
            request_id: { type: string }
  responses:
    Session:
      description: Spendly session
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Session' }
    Error:
      description: API error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }